+++ bin/typedef/linetv-channel-data.js
@@ -1,0 +1,11 @@
+/**
+ * @typedef LINETvChannelData
+ * @property {string} channelId
+ * @property {string} channelName
+ * @property {string} thumbnailUrl
+ * @property {string} serviceUrl
+ * @property {string|null} badgeType
+ * @property {string} channelEmblem
+ */
+"use strict";
+//# sourceMappingURL=linetv-channel-data.js.map