+++ bin/typedef/linetv-playlist-data.js
@@ -1,0 +1,8 @@
+/**
+ * @typedef LINETvPlaylistData
+ * @property {string} title
+ * @property {string} subtitle
+ * @property {Array<LINETvDetailPlaylistData>} playlists
+ */
+"use strict";
+//# sourceMappingURL=linetv-playlist-data.js.map