+++ CHANGELOG.md
@@ -5,6 +5,9 @@
  - `linetv list:modules` Gets a spotlight data unit list such as represent_clip, editor_pick_clip, and theme_playlist. 
  - `linetv get:spotlight` Gets spotlight data.
  - `linetv get:spotlight --format json` Gets spotlight data in JSON format.
+ - `linetv list:category` Gets a category list such as drama, music, etc.
+ - `linetv get:category` Gets category home data.
+ 
 ### Fixed
  - Fixed `npm test` failed on Windows
  - Fixed missing qs module